[SUGGESTION] You need to explain what the Modern::Perl module does.
And why you recommend using it, or otherwise doing what it does in your
programs.  You never say why every example program is run with use Modern::perl;
at the top, or what it does for you, or why you should use it, and if you
recommend actually adding it as a dependency to your readers CPAN modules, of if
they should just do what it does manually.