[SUGGESTION] I just call it CPAN, and drop the 'the.'
I guess this is really just a matter of preference, and maybe grammar or
semantics or something, but I call it CPAN as though CPAN is not an acronym, but
rather just another word.  I know CPAN is an acronym, but I never refer to it as
the Comprehensive Perl Archive Network.  It's just CPAN.