[SUGGESTION] Emboldening broken in html version.
Maybe it's a property of the programlisting tag that you use, but nothing is
emboldened on my html printout, which make this very confusing.  The B<> tag is
used in this source code, but it is not actually emboldened in the output from
Pod::PseudoPod.  Please fix this before it goes to the printers, because it may
not be bold in the real printed book version too.