[SUGGESTION] Why is it called "elsif" instead of "else if?"
Is this, because the parser is too stupid, or is there some other reason.  I
don't know if it'd be worth trying to find out though.  Because neither perlsyn
or the camel book explain why.

Also, both perlsyn and the camel book get away with out saying why, so you
probably can too.