[SUGGESTION] Should your output from your examples be listed in the book?
While reading I wanted to know what the output from the examples was, without
going over to my computer to try them.  Should each 'programlisting' have a
corresponding 'screen' section displaying the output from your example?  Or is
this a waste of paper and ink?

If you wanted to do this, you could probably write a fairly simple perl script
to parse out each 'programlisting' and add a 'screen' after it, or add another
tag, say 'programlistingwithexample' so that you could specify which examples
had their output listed, and which didn't.

I don't know if this is worth the effort, but consider it.