[INACCURATE] sidebar inaccurate about using each in for loops.
As shown in the program below similar to the one in your book, you actually can
use each in a for loop.  It just has to be a C style for (;;) loop, and not a
foreach loop called for.  This is a silly technicality that I'm not sure you
want to bother acknowledging, but I'm reporting it just to let you know.  I'll
leave it to yo to decide if this is worth clearing up.

%addresses = qw(a 1 b 2 c 3 d 4);
for (;my ($z, $y) = each %addresses;) {
print "$z lives at $y\n";
}

OUTPUT:
c lives at 3
a lives at 1
b lives at 2
d lives at 4