[SUGGESTION] You unreasonably suggest we watch our function calls if using each
As shown in the diff, you actually suggest (Okay maybe it was just implied, but
still.) that we check the functions we call to see if they use each, because
each's internal iterator is global, and it could mess with you iterating over
each.  This is unreasonable, and I doubt anyone does it.  Instead, this sounds
more like a bug in Perl 5 that should just be fixed.