Deleted a confusing section.
The contradiction the section this commit deletes really annoyed me, because it
never existed until you brought it up.  And the worst part is that it interested
me, and now I want to know the answer to a few questions the section posed.

Please either accept this commit and just delete this section, or further
explain it.  Maybe give example answers to the questions you ask.