[SUGGESTION] Delete confusing sidebar that makes me want to know more.
This sidebar is really anoying.  Now I want to know about this "deep magic" and
when it might be necessary, and how you use it.  But I doubt you want to bother
writing up a section explaining a little know and used "deep magic" feature.

Instead, could you please just delete this sidebar, because it's confusing as
is, and I doubt you'd want to bother writing up a brief section about a mostly
useless "deep magic" feature.