Deleted confusing, unneeded sidebar.
The sidebar deleted in this commit describes a depricated feature that I've
never heard of or come across.  I am definately a less experienced Perl
programmer than the author is, but I've never come across this depricated
feature.  Is it even in perl's own documenation in pre 5.10 versions?

I don't know, but I think you should delete it, because it's confusing and
unneeded.