[SUGGESTION] Are metacharacters atoms too?
You define atom as "each character," but what about multicharacter
metacharacters.  Are they atoms?  Is \s an atom?  Or is the \ and then the s
each their own atoms?

I know you haven't explain metacharacters yet, and there is not related
"metacharacters" section, which there probably should be, so I guess you
could create on if you think it's needed.

I know this occured in a section title "Literals," and as such probably only
refers to "literals," but this really did confuse me.