Deleted partial reference footnote to Jeffrey Freidl's book.
You have a footnote about seeing a book to further explain something without any
specific page reference in that book.  I think you should just delete the
footnote, because you don't want to try to find where that section in the book
is.  I briefly tried and failed, and I really do kinda want to read it, but I
can't, because I'm not going the read the whole book now, or spend hours looking
for that section.