Cleared up a confusing sentence.
You mentioning the "Perl 5 optree" makes me want to know more about the perl 5
optree, and how it all works.  Not only that, but I almost added a suggestion
for you to actually explain it, which isn't needed, so I think that part should
just be deleted to clear it up.